policymanager/requirements.txt
Flask
uWSGI==2.0.17.1
gitpython